# Gryffindor/openapi.yaml
...
swagger: "2.0"
info:
  description: Order Management API."
  title: "Gryffindor API"
  version: "1.0.0"
host: "gryffindor-1550651944789.appspot.com"
...
consumes:
- "application/json"
produces:
- "application/json"
schemes:
# Uncomment the next line if you configure SSL for this API.
#- "https"
- "http"
paths:
  "/gryffindor":
    post:
      description: "Gryffindor back api."
      operationId: "gryffindor"
      produces:
      - "application/json"
      responses:
        200:
          description: "Gryffindor"
          schema:
            $ref: "#/definitions/gryffindor"
      parameters:
      - description: "gryffindor"
        in: body
        name: message
        required: true
        schema:
          $ref: "#/definitions/gryffindor"
      security:
      - api_key: []
...
securityDefinitions:
  # This section configures basic authentication with an API key.
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"
  # This section configures authentication using Google API Service Accounts
  # to sign a json web token. This is mostly used for server-to-server
  # communication.
  google_jwt:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    # This must match the 'iss' field in the JWT.
    x-google-issuer: "jwt-client.endpoints.sample.google.com"
    # Update this with your service account's email address.
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/jwk/YOUR-SERVICE-ACCOUNT-EMAIL"
  # This section configures authentication using Google OAuth2 ID Tokens.
  # ID Tokens can be obtained using OAuth2 clients, and can be used to access
  # your API on behalf of a particular user.
  google_id_token:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://accounts.google.com"
    x-google-jwks_uri: "https://www.googleapis.com/oauth2/v1/certs"